ggEdit_Module	R Documentation

ggEdit

Description

The ggedit module.

Usage

ggEdit(
  input,
  output,
  session,
  obj,
  verbose = TRUE,
  showDefaults = FALSE,
  width = "auto",
  height = "auto"
)

Arguments

input
shinyapp input argument

output
shinyapp output argument

session
shinyapp session argument

obj
ggplot as reactive shiny object

verbose
logical to control if the output includes script for layers and themes calls for parsing to create objects (default, verbose=F)

showDefaults
toggle to control if the verbose output shows all the input arguments passed to the proto object (if verbose==FALSE then ignored)

width, height
Must be a valid CSS unit (like '100%', '400px', 'auto') or a number, which will be coerced to a string and have 'px' appended.