gggsave	R Documentation

gggsave

Description

Wrapper of ggsave that saves ggplot or list of ggplot objects to image or pdf.

Usage

gggsave(filename = "Rplot.pdf", plot = last_plot(), ...)

Arguments

filename
a character string giving the name of the file. If it is of the form "|cmd", the output is piped to the command given by cmd. If it is NULL, then no external file is created (effectively, no drawing occurs), but the device may still be queried (e.g., for size of text). For use with onefile = FALSE give a C integer format such as "Rplot%03d.pdf" (the default in that case)

plot
ggplot or list of ggplots to save, defaults to last plot displayed

...
other arguments passed on to graphics device

Details

default output is to create one pdf regardless of size of list of plots inputted

Value

nothing

Examples

## Not run: gggsave(pList)