plot.ggedit	R Documentation

Plots ggedit plot output objects

Description

Plots ggplot2 plots edited by ggedit with layout functionality.

Usage

## S3 method for class 'ggedit'
plot(x, ...)

Arguments

x
ggedit output of list of plots from the ggedit output

...
list which defines the location of each plot within the viewport layout. (see details)

Details

If ... is NULL then a default layout is used filling in a grid.layout row.wise. Defining plot.layout as a nested list giving the rows and the column indicies of each plot will produce a specific layout.

Examples

p=ggplot(iris,aes(x =Sepal.Length,y=Sepal.Width))
p=p+geom_point(aes(colour=Species))+geom_line()
p
x=ggedit(p)
plot(x)
plot(x$UpdatedPlots)
p1=pList[c(1:2)]
x1=ggedit(p1)
plot(x1)
plot(x1,plot.layout = list(list(rows=2,cols=2),list(rows=1,cols=1:2)))